+++ packages/adhere-ui-searchtable/es/Extension/DisabledOption/index.d.ts
@@ -1,0 +1,6 @@
+declare const _default: ({ className, style, children }: {
+    className: any;
+    style: any;
+    children: any;
+}) => JSX.Element;
+export default _default;