+++ packages/adhere-ui-searchtable/es/ProSearchTable.js
@@ -1,0 +1,2 @@
+import{__extends}from"tslib";import{MemoManager}from"./Extension/SearchAndPaginParams";import ProTableFactory from"./ProTableFactory";import SearchTableImplement from"./SearchTableImplement";var SearchAndPaginParamsMemo=MemoManager.create(),ProSearchTable=function(e){function a(){return null!==e&&e.apply(this,arguments)||this}return __extends(a,e),a}(SearchTableImplement);export default ProTableFactory(ProSearchTable,SearchAndPaginParamsMemo);export{SearchAndPaginParamsMemo};
+//# sourceMappingURL=ProSearchTable.js.map