+++ packages/adhere-ui-searchtable/src/ProSearchTable.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import { MemoManager } from './Extension/SearchAndPaginParams';
+import ProTableFactory from './ProTableFactory';
+import SearchTableImplement from './SearchTableImplement';
+
+export const SearchAndPaginParamsMemo: any = MemoManager.create();
+
+/**
+ * ProSearchTable
+ * @classdesc 高级查询表格
+ */
+class ProSearchTable extends SearchTableImplement {}
+
+export default ProTableFactory(ProSearchTable, SearchAndPaginParamsMemo);