+++ packages/adhere-ui-searchtable/src/index.ts
@@ -1,9 +1,21 @@
-import SearchTable from './searchtable';
-import SearchTableImplement from './searchtableimplement';
-import SearchTableStateImplement from './searchtablestateimplement';
+import DisabledOption from './Extension/DisabledOption';
+import LinkColumn from './Extension/LinkColumn';
+import OptionsWrap from './Extension/OptionsWrap';
+import * as SearchAndPaginParams from './Extension/SearchAndPaginParams';
+import ProSearchStateTable from './ProSearchStateTable';
+import ProSearchTable from './ProSearchTable';
+import SearchTableImplement from './SearchTableImplement';
+import SearchTableStateImplement from './SearchTableStateImplement';
+import SearchTable from './Searchtable';
 
 export default {
   Table: SearchTable,
   TableImplement: SearchTableImplement,
   TableStateImplement: SearchTableStateImplement,
+  ProSearchTable,
+  ProSearchStateTable,
+  DisabledOption,
+  LinkColumn,
+  OptionsWrap,
+  SearchAndPaginParams,
 };